Adds type and default value to parameter table for Kotlin source
Adds a lazy metadata node that we can look up when generating
the DAC parameter table so we can add type (with link support)
and default value information to the table.

Example table: (tab = space between columns)

Parameters
text: String	The text to display.
modifier: Modifier = Modifier.None	Modifier to be applied to the button.
onClick: () -> Unit = null	Will be called when the user clicks the button. The button will be disabled if it is null.
style: ButtonStyle = ContainedButtonStyle()	Contains the styling parameters for the button.

Bug: b/139411548
Bug: b/139407818
Bug: b/143536172
Test: ./gradlew core:cleanTest core:test
Change-Id: Ie62ab30ee18c082d2a486df644a1cfae8f54703b